requirements.txt
chalice==1.29.0
python-dotenv==1.0.0
marshmallow==3.20.1
pre-commit==3.4.0
pytest==7.4.2
pymongo==4.5.0
boto3==1.28.67
dependency-injector==4.41.0
